src/main/java/org/academiadecodigo/codezillas/portfolioApp/domainModel/skill/Skill.java
package org.academiadecodigo.codezillas.portfolioApp.domainModel.skill;

import org.academiadecodigo.codezillas.portfolioApp.domainModel.AbstractModel;
import org.academiadecodigo.codezillas.portfolioApp.domainModel.role.Role;

import javax.persistence.Entity;
import javax.persistence.ManyToMany;
import java.util.HashSet;
import java.util.Objects;
import java.util.Set;

@Entity
public class Skill extends AbstractModel {

    private String skillName;

    @ManyToMany(mappedBy = "skillSet")
    private final Set<Role> roleSet = new HashSet<>();

    public String getSkillName() {
        return skillName;
    }

    public void setSkillName(String skillName) {
        this.skillName = skillName;
    }

    public Set<Role> getRoleSet() {
        return roleSet;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Skill skill = (Skill) o;
        return Objects.equals(skillName, skill.skillName);
    }

    @Override
    public int hashCode() {
        return Objects.hash(skillName);
    }

    @Override
    public String toString() {
        return "Skill{" +
                "skillName='" + skillName + '\'' +
                "} " + super.toString();
    }
}
